@@ -9,18 +9,18 @@ url = https://github.com/Gradiant/pyodi
 
 [options]
 packages = find_namespace:
-python_requires = >=3.7
+python_requires = >=3.8
 install_requires =
-  numpy==1.20.1
-  loguru==0.5.3
-  matplotlib==3.3.4
-  numba>=0.52.0
-  pandas==1.2.1
-  pillow==8.1.0
-  plotly==5.2.2
-  pycocotools==2.0.2
+  numpy==1.22.1
+  loguru==0.6.0
+  matplotlib==3.5.2
+  numba>=0.56.0
+  pandas==1.4.2
+  pillow==9.2.0
+  plotly==5.9.0
+  pycocotools==2.0.4
   kaleido==v0.2.1
-  scikit-learn==0.24.1
+  scikit-learn==1.1.1
   fire==0.4.0
 
 [options.packages.find]
@@ -31,19 +31,19 @@ exclude =
 
 [options.extras_require]
 dev =
-  black==20.8b1
-  flake8==3.8.4
-  flake8-docstrings==1.5.0
-  isort==5.7.0
-  mkdocs==1.1.2
-  mkdocstrings==0.14.0
-  mkdocs-material==6.2.8
+  black==22.6.0
+  flake8==4.0.1 
+  flake8-docstrings==1.6.0
+  isort==5.10.1
+  mkdocs==1.3.1
+  mkdocstrings==0.19.0
+  mkdocs-material==8.3.9
   mock==4.0.3
-  mypy==0.800
-  pre-commit==2.10.1
-  pydocstyle==5.1.1
-  pymdown-extensions==8.1.1
-  pytest==6.2.2
+  mypy==0.960
+  pre-commit==2.20.0
+  pydocstyle==6.1.1
+  pymdown-extensions==9.5
+  pytest==7.1.2
 
 [options.entry_points]
 console_scripts =
